backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.28.35
botocore==1.31.35
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron_descriptor==1.4.0
cryptography==41.0.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.2
django-allauth==0.50.0
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==22.1
django-storages==1.13.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
future==0.18.3
geocoder==1.38.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.4
Markdown==3.4.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
Pillow==10.0.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
ratelim==0.1.6
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
scipy==1.13.0
seaborn==0.13.2
sendgrid==6.11.0
sentry-sdk==1.27.1
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==5.4.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.1.42
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.12
zipp==3.15.0
